Sistemata visualizzazione e logica di reset dei timer dei led

+++ MQTT.Sharing/Models/PublisherPayload.cs
@@ -20,6 +20,6 @@ namespace MQTT.Sharing.Models
         public string Timestamp { get; set; }
 
         [JsonPropertyName("battery")]
-        public int Battery { get; set; }
+        public decimal Battery { get; set; }
     }
 }
